修改资源id字符串转int方法，支持大于0x7f的情况

+++ FairyPlugin/src/main/java/com/limpoxe/fairy/util/ResourceUtil.java
@@ -88,7 +88,7 @@ public class ResourceUtil {
         }
         if (idHex != null) {
             try {
-                int id = Integer.parseInt(idHex, 16);
+                int id = (int)Long.parseLong(idHex, 16);
                 return id;
             } catch (Exception e) {
                 LogUtil.printException("ResourceUtil.parseResId", e);
